user details for permisssion based display

+++ src/Modules/CommitDetails/Views/CommitDetailsHTMLView.tpl.php
@@ -3,17 +3,13 @@
         <div class="navbar-default col-sm-2 sidebar" role="navigation">
 		<div class="sidebar-nav ">
 			<ul class="nav in" id="side-menu">
-				<li class="sidebar-search">
-					<div class="input-group custom-search-form hvr-bounce-in"">
-						<input type="text" class="form-control" placeholder="Search...">
-						<span class="input-group-btn">
-							<button class="btn btn-default" type="button">
-								<i class="fa fa-search"></i>
-							</button>
-                        </span>
-					</div>
-                    <!-- /input-group -->
-                </li>
+
+
+                <?php
+
+                if (in_array($pageVars["data"]["current_user_role"], array("1", "2"))) {
+
+                ?>
                 <li>
                     <a href="/index.php?control=Index&amp;action=show" class="hvr-bounce-in">
                         <i class="fa fa-dashboard fa-fw hvr-bounce-in"></i> Dashboard
@@ -29,7 +25,12 @@
                         <i class="fa  fa-cog fa-fw hvr-bounce-in"></i> Configure
                     </a>
                 </li>
-                
+
+                <?php
+
+                }
+                ?>
+
                 <li>
                     <a href="index.php?control=FileBrowser&action=show&item=<?php echo $pageVars["data"]["repository"]["project-slug"] ; ?>" class="hvr-bounce-in">
                         <i class="fa fa-folder-open-o hvr-bounce-in"></i> File Browser
@@ -45,11 +46,24 @@
                         <i class="fa fa-history fa-fw hvr-bounce-in""></i> History <span class="badge"></span>
                     </a>
                 </li>
-                <li>
-                    <a href="index.php?control=CommitDetails&action=delete&item=<?php echo $pageVars["data"]["repository"]["project-slug"] ; ?>" class="hvr-bounce-in">
-                        <i class="fa fa-trash fa-fw hvr-bounce-in""></i> Delete
-                    </a>
-                </li>
+
+                <?php
+
+                if (in_array($pageVars["data"]["current_user_role"], array("1", "2"))) {
+
+                    ?>
+
+                    <li>
+                        <a href="index.php?control=CommitDetails&action=delete&item=<?php echo $pageVars["data"]["repository"]["project-slug"] ; ?>" class="hvr-bounce-in">
+                            <i class="fa fa-trash fa-fw hvr-bounce-in""></i> Delete
+                        </a>
+                    </li>
+
+                <?php
+
+                }
+                ?>
+
             </ul>
         </div>
         </div>
